+++ ios/RNPlayerView.swift
@@ -108,6 +108,7 @@ public class RNPlayerView: ExpoView {
     let onBmpCueExit = EventDispatcher()
     let onBmpMetadata = EventDispatcher()
     let onBmpMetadataParsed = EventDispatcher()
+    let onBmpFairplayLicenseAcquired = EventDispatcher()
 
     required init(appContext: AppContext? = nil) {
         super.init(appContext: appContext)
@@ -510,6 +511,10 @@ extension RNPlayerView: PlayerListener {
 
     public func onCastWaiting(forDevice event: CastWaitingForDeviceEvent, player: Player) {
         onBmpCastWaitingForDevice(event.eventPayload())
+    }
+
+    public func onFairplayLicenseAcquired(_ event: FairplayLicenseAcquiredEvent, player: Player) {
+        onBmpFairplayLicenseAcquired(event.eventPayload())
     }
 #endif
 }