+++ src/components/PlayerView/index.tsx
@@ -206,6 +206,7 @@ export function PlayerView({
         props.onVideoPlaybackQualityChanged
       )}
       onBmpDownloadFinished={proxy(props.onDownloadFinished)}
+      onBmpFairplayLicenseAcquired={proxy(props.onFairplayLicenseAcquired)}
     />
   );
 }